requirements.txt
nltk
PyPDF2
pytesseract
pdf2image
openai
google-generativeai
flask
flask-sqlalchemy
flask-login
flask-socketio
flask-cors
python-dotenv
werkzeug
eventlet
